- fix(types): 统一导入到 src/types.ts，移除 shared/types 残留路径 - chore(tsconfig): 将 include 扩展为 src/**/* 覆盖迁移后的源文件 - feat(build): Vite 设置 root 为 src，并将 build.outDir 设为 ../dist 以匹配 Tauri frontendDist - refactor(api): 去除未使用的 plugin-shell import；统一 electronAPI 类型定义至 vite-env.d.ts - build: 验证 renderer 构建通过，产物输出至 dist/
